Working major on a firebase based Web service.
Card authentication will be delegated to firebase and cloud functions.
The card authentication key will be stored on cloud, as this key should be
kept secure but not confidential.

The card authentication key is used to establish an encrypted channel between
card terminal program and chip. Communications on this channel is not revealed
to Internet, so there's no hazard to obtain an encryption key from web service.